feat: make material type pipe, delete useless function from template
