+++ libs/users/materials/feature-materials/src/lib/material-content/pipes/MaterialType.pipe.spec.ts
@@ -1,0 +1,11 @@
+/* tslint:disable:no-unused-variable */
+
+import { TestBed, async } from '@angular/core/testing';
+import { MaterialTypePipe } from './MaterialType.pipe';
+
+describe('Pipe: MaterialTypee', () => {
+  it('create an instance', () => {
+    let pipe = new MaterialTypePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
