+++ libs/users/materials/feature-materials/src/lib/material-content/pipes/MaterialType.pipe.ts
@@ -1,0 +1,22 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'MaterialType',
+  standalone:true
+})
+export class MaterialTypePipe implements PipeTransform {
+
+  transform(link: string): string {
+    const extension = link.split('.').pop();
+  
+    switch(extension) {
+      case 'pdf':
+        return 'pdf';
+      case 'mp3':
+        return 'audio';
+      default:
+        return 'video';
+    }
+  }
+
+}
